// app/models/graphql.tsx
export type Maybe<T> = T | null;
export type InputMaybe<T> = Maybe<T>;
export type Exact<T extends { [key: string]: unknown }> = {
  [K in keyof T]: T[K];
};
export type MakeOptional<T, K extends keyof T> = Omit<T, K> & {
  [SubKey in K]?: Maybe<T[SubKey]>;
};
export type MakeMaybe<T, K extends keyof T> = Omit<T, K> & {
  [SubKey in K]: Maybe<T[SubKey]>;
};
/** All built-in and custom scalars, mapped to their actual values */
export type Scalars = {
  ID: string;
  String: string;
  Boolean: boolean;
  Int: number;
  Float: number;
  ConnectionCursor: any;
  ConnectionLimitInt: any;
  Date: any;
  DateTime: any;
  Email: any;
  JSONObject: any;
};
// ...
export type CatalogBooleanFilter = {
  /** The name of the filter */
  name: CatalogBooleanFilterName;
  /** The filter value */
  value: Scalars['Boolean'];
};

/** The list of currently supported top level Catalog props on which catalog items can be filtered. */
export enum CatalogBooleanFilterName {
  /** isBackorder */
  IsBackorder = 'isBackorder',
  /** isDeleted */
  IsDeleted = 'isDeleted',
  /** isLowQuantity */
  IsLowQuantity = 'isLowQuantity',
  /** isSoldOut */
  IsSoldOut = 'isSoldOut',
  /** isVisible */
  IsVisible = 'isVisible',
}

/** Catalog items are combined to create a catalog. Each item can represent a different type of content. */
export type CatalogItem = {
  /** Item ID */
  _id: Scalars['ID'];
  /** Date and time at which the item was created */
  createdAt?: Maybe<Scalars['DateTime']>;
  /** Shop that owns the item */
  shop: Shop;
  /** Date and time at which the item was last updated */
  updatedAt?: Maybe<Scalars['DateTime']>;
};
// ...
export type CatalogItemConnection = {
  __typename?: 'CatalogItemConnection';
  /** The list of nodes that match the query, wrapped in an edge to provide a cursor string for each */
  edges?: Maybe<Array<Maybe<CatalogItemEdge>>>;
  /**
   * You can request the `nodes` directly to avoid the extra wrapping that `NodeEdge` has,
   * if you know you will not need to paginate the results.
   */
  nodes?: Maybe<Array<Maybe<CatalogItem>>>;
  /** Information to help a client request the next or previous page */
  pageInfo: PageInfo;
  /** The total number of nodes that match your query */
  totalCount: Scalars['Int'];
};
// ...
export type CatalogItemEdge = {
  __typename?: 'CatalogItemEdge';
  /** The cursor that represents this node in the paginated results */
  cursor: Scalars['ConnectionCursor'];
  /** The catalog item */
  node?: Maybe<CatalogItem>;
};
// ...
export enum CatalogItemSortByField {
  /** Sort by item ID */
  Id = '_id',
  /** Sort by date and time at which the item was created */
  CreatedAt = 'createdAt',
  /** Sort in the shop-defined order for the tag filter */
  Featured = 'featured',
  /** Sort by price */
  MinPrice = 'minPrice',
  /** Sort by date and time at which the item was last updated */
  UpdatedAt = 'updatedAt',
}
// ...
export type PageInfo = {
  __typename?: 'PageInfo';
  /** When paginating forwards, the cursor to continue. */
  endCursor?: Maybe<Scalars['ConnectionCursor']>;
  /** When paginating forwards, are there more items? */
  hasNextPage: Scalars['Boolean'];
  /** When paginating backwards, are there more items? */
  hasPreviousPage: Scalars['Boolean'];
  /** When paginating backwards, the cursor to continue. */
  startCursor?: Maybe<Scalars['ConnectionCursor']>;
};
// ...
export type QueryCatalogItemsArgs = {
  after?: InputMaybe<Scalars['ConnectionCursor']>;
  before?: InputMaybe<Scalars['ConnectionCursor']>;
  booleanFilters?: InputMaybe<Array<InputMaybe<CatalogBooleanFilter>>>;
  first?: InputMaybe<Scalars['ConnectionLimitInt']>;
  last?: InputMaybe<Scalars['ConnectionLimitInt']>;
  offset?: InputMaybe<Scalars['Int']>;
  searchQuery?: InputMaybe<Scalars['String']>;
  shopIds: Array<InputMaybe<Scalars['ID']>>;
  sortBy?: InputMaybe<CatalogItemSortByField>;
  sortByPriceCurrencyCode?: InputMaybe<Scalars['String']>;
  sortOrder?: InputMaybe<SortOrder>;
  tagIds?: InputMaybe<Array<InputMaybe<Scalars['ID']>>>;
};
// ...
export type Shop = Node & {
  __typename?: 'Shop';
  /** The shop ID */
  _id: Scalars['ID'];
  /** An the shop's default address */
  addressBook?: Maybe<Array<Maybe<Address>>>;
  /** Whether to allow user to checkout without creating an account */
  allowGuestCheckout?: Maybe<Scalars['Boolean']>;
  /** The base unit of length */
  baseUOL?: Maybe<Scalars['String']>;
  /** The base unit of Measure */
  baseUOM?: Maybe<Scalars['String']>;
  /** URLs for various shop assets in various sizes */
  brandAssets?: Maybe<ShopBrandAssets>;
  /** The default shop currency */
  currency: Currency;
  /** The default navigation tree for this shop */
  defaultNavigationTree?: Maybe<NavigationTree>;
  /** The ID of the shop's default navigation tree */
  defaultNavigationTreeId?: Maybe<Scalars['String']>;
  /** Default parcel size for this shop */
  defaultParcelSize?: Maybe<ShopParcelSize>;
  /** Shop description */
  description?: Maybe<Scalars['String']>;
  /** The shop's default email record */
  emails?: Maybe<Array<Maybe<EmailRecord>>>;
  /** Returns a list of groups for this shop, as a Relay-compatible connection. */
  groups?: Maybe<GroupConnection>;
  /** Shop's keywords */
  keywords?: Maybe<Scalars['String']>;
  /** Shop default language */
  language: Scalars['String'];
  /** Shop name */
  name: Scalars['String'];
  /** Returns a list of roles for this shop, as a Relay-compatible connection. */
  roles?: Maybe<RoleConnection>;
  /** Returns URLs for shop logos */
  shopLogoUrls?: Maybe<ShopLogoUrls>;
  /** Shop's type */
  shopType?: Maybe<Scalars['String']>;
  /** Shop's slug */
  slug?: Maybe<Scalars['String']>;
  /** Returns URLs for various storefront routes */
  storefrontUrls?: Maybe<StorefrontUrls>;
  /** Returns a paged list of tags for this shop */
  tags?: Maybe<TagConnection>;
  /** Shop default timezone */
  timezone?: Maybe<Scalars['String']>;
  /** The shop's units of length */
  unitsOfLength?: Maybe<Array<Maybe<UnitOfLength>>>;
  /** The shop's units of measure */
  unitsOfMeasure?: Maybe<Array<Maybe<UnitOfMeasure>>>;
};
// ...
export enum SortOrder {
  /** ascending */
  Asc = 'asc',
  /** descending */
  Desc = 'desc',
}